[12.x] Introduce Arr::from()
